test(administration): split extension tooling e2e spec below the line limit
e2e.spec.ts crossed the 500-line ADR threshold; the full-project and
scaffolded-configs suites move into e2e.spec/ per
adr/2026-05-06-split-large-administration-test-files.md. As a bonus the
two suites now run in parallel jest workers.
